BehaviourGuard/requirements.txt
#Copyright (c) 2025 KOMAL MAZHAR MUSHTAQ All rights reserved
# Core Flask Framework
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Task Queue and Background Processing
Celery==5.5.3
redis==6.4.0

# HTTP Requests and API Integration
requests==2.32.4

# File Processing and Analysis
python-magic-bin==0.4.14
yara-python==4.5.4
psutil==5.9.0

# Report Generation
reportlab==4.4.3
matplotlib==3.10.5
numpy==1.26.4

# Data Validation
email-validator==2.2.0

# Production Server
gunicorn==23.0.0

# Development and Testing (optional)
pytest==7.0.0
pytest-flask==1.2.0
black==23.0.0
flake8==6.0.0

# Additional Dependencies
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.7.0
kombu==5.3.4
billiard==4.2.0
amqp==5.2.0
vine==5.1.0
urllib3==2.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
packaging==23.2
pyparsing==3.1.1
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
pillow==10.2.0
python-dateutil==2.8.2
six==1.16.0
contourpy==1.2.0
kiwisolver==1.4.5
